+++ lib/extensions/string_extensions.dart
@@ -1,0 +1,3 @@
+extension Util on String {
+  String toBold() => "**$this**";
+}